api_gateway/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-pubsub
python-multipart